requirements.txt
# requirements.txt
argostranslate==1.9.6
python-docx==1.1.2
tqdm==4.67.1
pyyaml